scaler-course
# Conflicts:
#	.idea/compiler.xml
#	pom.xml
#	src/main/java/com/scaler/DesignPattern/DesignPatternNotes
#	src/main/java/com/scaler/LLDProject/googledrive/FileSystemComponent.java
#	src/main/java/com/scaler/LLDProject/googledrive/FileSystemProxy.java
#	src/main/java/com/scaler/LLDProject/googledrive/Folder.java
#	src/main/java/com/scaler/LLDProject/googledrive/Main.java
#	src/main/java/com/scaler/LLDProject/googledrive/PermissionsDecorator.java
#	src/main/java/com/scaler/MainApplication.java
#	src/main/java/com/scaler/MainTest.java
#	src/main/java/com/scaler/Scaler/Graph/Algorithm/BellmannFord.java
#	src/main/java/com/scaler/Scaler/Graph/Algorithm/Prims_MST.java
#	src/main/java/com/scaler/Scaler/Graph/Capture_Regions_on_Board.java
#	src/main/java/com/scaler/Scaler/Graph/CheckCycle_UnDirected_Graph_Using_DSU.java
#	src/main/java/com/scaler/Scaler/Graph/CycleIn_DirectedGraph.java
#	src/main/java/com/scaler/Scaler/Graph/Cycle_IN_UnDirected_Graph.java
#	src/main/java/com/scaler/Scaler/Graph/Minimum_Weighted_Cycle.java
#	src/main/java/com/scaler/Scaler/MathDSA/GeneralProblem/SquareRootOfNumber.java
#	src/main/java/com/scaler/Scaler/MathDSA/PrimeNumber/PrimeNumber.java
#	src/main/java/com/scaler/Scaler/Sorting/InversionCount.java
#	src/main/java/com/scaler/Solution.java

+++ src/main/java/com/scaler/Solution.java
@@ -290,7 +290,7 @@ class Solution3 {
         for (int i = M - 1; i >= 0; i--) {
             sum = carry;
             for (int j = 0; j < A; j++) {
-                sum += (int) (C[j].charAt(i) - '0');
+                sum += (C[j].charAt(i) - '0');
             }
 
             ans.append((char) ((char) (sum % 2) + '0'));
